rename: 패키지 정리
- 도메인 별로 묶었음
