--- src/main/java/finalmission/domain/Member.java
@@ -1,9 +1,0 @@
-package finalmission.domain;
-
-import finalmission.login.MemberType;
-
-public interface Member {
-
-    Long getId();
-    MemberType getType();
-}
