+++ src/main/java/finalmission/domain/login/MemberType.java
@@ -1,4 +1,4 @@
-package finalmission.login;
+package finalmission.domain.login;
 
 public enum MemberType {
     CREW, COACH
