+++ src/main/java/finalmission/domain/member/Member.java
@@ -1,0 +1,9 @@
+package finalmission.domain.member;
+
+import finalmission.domain.login.MemberType;
+
+public interface Member {
+
+    Long getId();
+    MemberType getType();
+}
