+++ src/main/java/finalmission/repository/CrewRepository.java
@@ -1,7 +1,6 @@
 package finalmission.repository;
 
-import finalmission.domain.Crew;
-import java.util.List;
+import finalmission.domain.member.Crew;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
