+++ src/main/java/finalmission/repository/ReservationRepository.java
@@ -1,6 +1,6 @@
 package finalmission.repository;
 
-import finalmission.domain.Reservation;
+import finalmission.domain.reservation.Reservation;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
