+++ src/main/java/finalmission/repository/ReservationTimeRepository.java
@@ -1,6 +1,6 @@
 package finalmission.repository;
 
-import finalmission.domain.ReservationTime;
+import finalmission.domain.reservation.ReservationTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationTimeRepository extends JpaRepository<ReservationTime, Long> {
