+++ src/test/java/finalmission/service/ReservationServiceTest.java
@@ -2,7 +2,7 @@ package finalmission.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import finalmission.domain.Reservation;
+import finalmission.domain.reservation.Reservation;
 import finalmission.dto.ReservationRequestDto;
 import finalmission.dto.ReservationResponse;
 import java.time.LocalDate;
